% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_git.R
\name{add_gitkeep_to_empty_dirs}
\alias{add_gitkeep_to_empty_dirs}
\title{Add .gitkeep to empty directories}
\usage{
add_gitkeep_to_empty_dirs(dir_path)
}
\arguments{
\item{dir_path}{The path to the directory to check for empty subdirectories.}
}
\description{
Add .gitkeep to empty directories
}
\keyword{internal}
