% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_git.R
\name{create_description}
\alias{create_description}
\title{Create DESCRIPTION file}
\usage{
create_description(project_name)
}
\arguments{
\item{project_name}{The name of the project.}
}
\description{
Create DESCRIPTION file
}
\keyword{internal}
