% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_git.R
\name{create_readme}
\alias{create_readme}
\title{Create README.md file}
\usage{
create_readme(project_name)
}
\arguments{
\item{project_name}{The name of the project.}
}
\description{
Create README.md file
}
\keyword{internal}
