% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_git.R
\name{template_git}
\alias{template_git}
\title{Create a psychology research project template}
\usage{
template_git(project_name, github = TRUE)
}
\arguments{
\item{project_name}{The name of the project to create.}

\item{github}{Logical, whether to create a GitHub repository and push the project.}
}
\description{
This function creates a new psychology research project with a predefined structure,
including necessary files and folders. It can also initialize a Git repository
and push the project to GitHub.
}
